backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
sqlalchemy>=1.4.0
pydantic>=1.10.0
pydantic[email]
PyJWT>=2.6.0
passlib[bcrypt]>=1.7.4
python-dotenv>=0.20.0
psycopg2-binary>=2.9.0
firebase-admin>=6.0.0
requests>=2.28.0
python-multipart>=0.0.5 